% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mmseqs.R
\name{load_mmseqs}
\alias{load_mmseqs}
\title{Load MMseqs}
\usage{
load_mmseqs(
  file,
  frag_thr = 3,
  sup_thr = 0.7,
  tax_ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  include_name = FALSE
)
}
\arguments{
\item{file}{The file to load.}

\item{frag_thr}{The threshold for total_fragments.}

\item{sup_thr}{The threshold for support.}

\item{tax_ranks}{The taxonomy ranks to include.}

\item{include_name}{Whether to include the sample name in the output.}
}
\value{
A data frame.
}
\description{
This function loads MMseqs data from a file, applies some filtering and transformations, and returns a data frame.
}
